Add a fully functional constraint solver.
This was developed in the "numerical" branch, but I'm merging in the
code by hand and throwing away the history (lots of stupid commits). So
sue me.

This constraint solver is based on gradient descent, with a small
regularization function to discourage needless motion. An obvious
direction for future work is adding a more sophisticated objective
function, for better user intuitiveness.

We presently support angles, coradial, and dimension constraints.